clean up variadic source generator
