--- src/Xabbo.Common.Generator/Internal/CS0518.Fix.cs
@@ -1,2 +1,0 @@
-namespace System.Runtime.CompilerServices;
-internal static class IsExternalInit;
