+++ src/Xabbo.Common.Generator/Model/VariadicType.cs
@@ -1,19 +1,11 @@
 namespace Xabbo.Common.Generator.Model;
 
 internal sealed record VariadicType(
-    string Namespace,
-    string Name,
+    string FullyQualifiedName,
     bool IsParser,
     bool IsComposer,
-    bool IsArray,
-    bool IsTypeKnown
+    bool IsArray
 )
 {
-    public override string ToString()
-    {
-        if (!string.IsNullOrWhiteSpace(Namespace))
-            return $"{Namespace}.{Name}";
-        else
-            return $"{Name}";
-    }
+    public override string ToString() => FullyQualifiedName;
 }
